Update value rule for omit type when performing group search in query window.
the type of literals in a group search can be automatically determined by its suffix or prefix or specific characters as follows(case insensitive):
the literal with the u or ul, lu suffix is of type ulong(8 Bytes),
the literal with the f suffix is of type float,
the literal with the d suffix is of type double,
the literal with the 0x prefix is of type Hex(hexadecimal),
the literal contains a decimal point (.) character is of type float,
the literal contains a~f characters is of type Hex(hexadecimal) when not parsed as float and double,
other literals of unspecified types, default to uint(4 Bytes) type.

Hex values can now have spaces or delimiters when searching for Hex types in query window.

______________

Added row number to Cheat list in main window.

Added support for loading cheat codes in batchcode format made by zsword in main window.
Currently, only the offset, value, size, and vtype fields are supported for parsing, and other fields are ignored.

batchcode cheat codes Example:
@batchcode|data|0|0|code||offset:0x7777777 value:0x0123456789ABCDEF size:8;offset:0xAABBCC value:0x0123456789 size:5|0|DescForBatchcode

______________

Added to confirm whether the PS4 Socket is disconnected and will try to reconnect automatically in PS4Tool.