Improve query and pointerFinder
add new options

QueryBufferSize:
Set the minimum buffer size (in MB) in querying and pointerFinder, enter 0 to not use buffer,
Setting this value to 0 is better when the total number of Sections in the game is low.
If the game has more than a thousand Sections, Buffer must be set.