New feature: undo scan and auto-pause when scanning.
Support redo scan(revert to the previous scan result), need to enable `EnableUndoScan` in the options.
Support automatically pause the game when starting the scan in query, need to enable `EnableScanAutoPause` in the options.
Support automatically resume the game when the scan is complete in query, need to enable `EnableScanDoneResume` in the options.

add new options:

EnableUndoScan: Determines whether to enable undo scan(revert to the previous scan result), if enabled, more memory needs to be used during scanning. Default enable.
EnableScanAutoPause: Determines whether to automatically pause the game when starting the scan in query. Default disabled.
EnableScanDoneResume: Determines whether to automatically resume the game when the scan is complete in query. Default disabled.