Added several experimental features as follows
1. Added collapsible split container ui in Query and HexEditor and PointerFinder (can be set in options to enable or not).

2. Added SimpleValue CheckBox in the query window. When the scan type is float or double, this feature can be enabled to reduce unnecessary results.

3. Added Not CheckBox in the query window. When this feature is enabled, the scanning logic can be reversed. For example, if the compare type is exact and Not is enabled, it becomes not exact.

4. Added AutoPause and DoneResume CheckBox in the query window, you can determine whether to enable or not.

5. Added multiple options parameters as follows

EnableCollapsibleContainer:
Determines whether to enable collapsible split container ui in Query and HexEditor and PointerFinder windows. Default enabled.

EnableCheatLock:
Determines whether to enable cheat lock in the main window. Default enabled.

FloatingSimpleValueExponents:
Determine the exponents value of the simple value of floating.
This option value is used in the query window when SimpleValues for floating point numbers is enabled.
Cheat Engine is set to 11 (2 to the 11th power = 2^11 = plus or minus 2048). Default value is 11.

EnableAutoRefresh:
Determines whether to enable Auto Refresh in HexEditor. Default disabled.

AutoRefreshTimerInterval:
Determines the Interval of AutoRefreshTimer when AutoRefresh is enabled, in milliseconds, HexEditor needs to be restarted after this value is changed. Default 2500.