Added AutoFindClosestChangedPosition Option for Hex Editor
Added support for automatic selection of the value change position closest to the current location (red and green value positions) when there is no input value for the find value, when the option "AutoFindClosestChangedPosition" is enabled (enabled by default).

Change the input from one value to two values (pause and resume intervals) when executing SlowMotion.