Fixed the issue of ScanTypeBox not restoring to the default state
after clicking the New button in the Query window.

[experimental]
In the SectionView of the Query window, the "Import" right-click option has been enabled. This option allows you to import dumped bin files into the specified PS4 section memory block.
Note that this functionality is experimental; after importing dump files, the game is likely to be abnormal or encounter errors around 99% of the time.