Implemented the LittleEndian endianness option in the Hex Editor window.
Previously, when searching with the input format set to Hex, the tool would use BigEndian to query the target. Now, an option has been added to allow users to decide whether to switch to LittleEndian for searching.
Note: When LittleEndian is enabled, regardless of input length, the input Hex values will be completely reversed (using Array.Reverse).

Added "InputIsHexFormat" in the Option window of the HexEditor.
Determines whether to enable the input format for searching as Hex in the HexEditor window. Default enabled.

Added "UsingLittleEndian" in the Option window of the HexEditor.
Determines whether to enable using LittleEndian endianness, when the input format for the search is Hex, in the HexEditor window, if this option is not checked, it will use BigEndian. Default disabled.