CheatGridView has been changed to VirtualMode, improving processing speed.
Fixed an issue that occurred when searching in the Query window and then refreshing, which caused abnormal behavior.

Added the CheatGridViewGroupByEnabled option, which allows you to enable or disable the grouping feature in CheatGridView.
Determines whether to enable the GroupByEnabled feature in CheatGridView. When enabled, Cheat items will be grouped. If you have a large number of Cheat items, you may consider disabling this feature to speed up Cheat loading. After setting this, you need to restart the program for it to take effect. Default enabled.