Added support for automatically detecting hidden sections.
When the "SectionViewDetectHiddenSection" option is enabled, SectionView will automatically force the display of sections that are hidden due to undefined addresses by the system. This feature is currently experimental.

Automatically detecting hidden sections is done by arranging the existing sections from the smallest Start address to the largest. If the memory addresses of two sections are not contiguous, a new section is automatically added to fill the addresses that are undefined by the system. This detection process is executed when the SectionViewDetectHiddenSection option is enabled in the settings.

Added SectionViewDetectHiddenSection, LastHiddenSectionLengthHex, and QuerySectionViewHiddenForeColor options in the Option window.