Added support for "Jump to Address" in the HexEditor's context menu.
Enter an address, and it will automatically navigate to that location.

Added support for "Jump to Offset" in the HexEditor's context menu. Enter an offset value from the current address, and it will navigate to the address relative to the current position. Negative offsets are also accepted.

Added support for setting the starting address of HiddenSection as the end address of the previous section. avoiding starting addresses that do not end with an integer 0. You can enable/disable this by setting the `HiddenSectionStartAtPreviousEnd` option, which is enabled by default.