Improved the stop mechanism for locking and refreshing cheats in  the Cheat window.
When unchecked LockEnable in the Cheat window, it will immediately stop the Task that locks cheats.
When unchecked AutoRefresh in the Cheat window, it will immediately stop the Task that refreshes cheats.
When double-clicking the Refresh button in the Cheat window, it will immediately stop the Task that refreshes cheats.