Added Filter CheckBox in the Min and Max Address fields of the Query window
Whether selected or not, the search will be limited by the input Minimum and Maximum address range for query targets. If selected, it will also filter the SectionView content based on the input Minimum and Maximum address range.

When performing a "SCAN for Hidden Sections" search in the Query window, the Prot of executable sections will automatically be set to 7.