Added support for "Invert checked Hidden" and "Disable checked Hidden"
in the right-click context menu of the SectionView in the Query window.

Disable checked Hidden:
Executing "Disable checked Hidden" will set the Valid of the checked Hidden sections to Disable, and update the settings written to the Hidden configuration file at the "path to PS4CheaterNeo\sections\[GAME_ID].conf" path.
To revert, manual editing of the conf file is required, changing the Valid of the Hidden sections from False to True.
Note: The Query window needs to Refresh Processes to incorporate the new Hidden conf settings.